requirements.txt
# Main project dependencies for all services
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
kafka-python==2.0.2
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Development dependencies
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1